piplineScrapingTwitter/requirements.txt
selenium==4.6.0
webdriver_manager==3.5.2
apache-airflow
webdriver_manager==3.5.2